Database logic moved into persistance layer.
